Database/requirements.txt
lxml
requests
pandas
numpy
scipy
opencv-python
tensorflow
google.cloud